fixed tests

+++ services/parser/widgets/project/db.py
@@ -22,10 +22,17 @@ job = Table(
     Column('link', String, unique=True)
 )
 
+category = Table(
+    'category', meta,
+    Column('id', Integer, primary_key=True),
 
-def create_tables():
+)
+
+
+def create_tables(drop_all: bool = False):
     engine = sa.create_engine(DSN.get())
-    meta.drop_all(bind=engine, tables=[job, user], checkfirst=True)
+    if drop_all:
+        meta.drop_all(bind=engine, tables=[job, user], checkfirst=True)
     meta.create_all(bind=engine, tables=[job, user], checkfirst=True)
     return engine
